Trim version string before comparing (#12)
The BUNDLE-VERSION file is usually created with a newline character at
the end if the file is manually created by the user. For e.g. when
language grammars are built manually. This causes the version comparison
to fail because "v0.10.0" != "v0.10.0\n". This fix trims the version
string before comparing.