Highlight ruby block delimiters consistently (#62)
There are two syntaxes for defining "blocks" in ruby. One is to use "do"
and "end" to mark the start and end of a block, the other is to use "{"
and "}". These are used almost identically and so they should be
highlighted in the same way.

Plus, braces are also used to denote a Hash literal. For example, `{ a:
1 }` is a Hash with a key of `:a` and a value of 1, while `{ :a }` is a
block that just yields `:a`. This change allows those two different use
cases to be visually distinguished.