Add syntax highlighting for YAML
The highlighting queries use capture names that correspond to
yaml-mode's font-lock faces. The result is almost identical, some small
differences exist due to the way the Tree-sitter parser for YAML is
implemented. For example, `block_scalar` includes the block scalar
header (block style indicator, block chomping indicator) which means
that the block scalar header is highlighted with the same face as the
actual block scalar (i.e. as a string), whereas in plain yaml-mode the
block scalar header does not have a font-lock face at all.