copy queries/ocaml to queries/ocaml-interface
the ML and MLI files have separate parsers, but appear to share the same
highlights.scm file. this change duplicates the file so that emacs can
find the queries when using the ocaml-interface grammar. this is
necessary to get proper highlighting in interface files. without it,
tree-sitter-hl-mode will silently fail to activate in mli files if the
language is set to ocaml-interface.