fix: address lint warnings detected by ts_query_ls
This commit address various lints like duplicate alternants in
alternations, superfluous escape chars (`\`), and unused underscored
captures.

Please let me know if you'd like a workflow to detect this in CI, and/or
one to ensure queries have been formatted with `ts_query_ls`. I'd be
happy to do this.